build: Testes unitários para os Api/Resources

+++ src/main/java/br/com/sevenfood/product/sevenfoodproductapi/core/service/ProductCategoryService.java
@@ -1,14 +1,17 @@
 package br.com.sevenfood.product.sevenfoodproductapi.core.service;
 
+import br.com.sevenfood.product.sevenfoodproductapi.application.api.exception.ResourceFoundException;
 import br.com.sevenfood.product.sevenfoodproductapi.core.domain.ProductCategory;
 import br.com.sevenfood.product.sevenfoodproductapi.core.ports.in.productcategory.*;
 import br.com.sevenfood.product.sevenfoodproductapi.core.ports.out.ProductCategoryRepositoryPort;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Slf4j
 @Service
 @AllArgsConstructor(onConstructor = @__(@Autowired))
 public class ProductCategoryService implements CreateProductCategoryPort, UpdateProductCategoryPort, FindByIdProductCategoryPort, FindProductCategoriesPort, DeleteProductCategoryPort {
@@ -45,9 +48,15 @@ public class ProductCategoryService implements CreateProductCategoryPort, Update
     @Override
     public boolean remove(Long id) {
         try {
+            ProductCategory productCategoryById = findById(id);
+            if (productCategoryById == null) {
+                throw new ResourceFoundException("Product Category not found");
+            }
+
             productCategoryRepository.remove(id);
             return Boolean.TRUE;
-        } catch (Exception e) {
+        } catch (ResourceFoundException e) {
+            log.error("Erro ao remover produto: {}", e.getMessage());
             return Boolean.FALSE;
         }
     }
